count

+++ GOBI_Block/src/dennis/counter/CounterUtils.java
@@ -13,7 +13,7 @@ public class CounterUtils {
 
 	// doubles werden abgerundet
 	public static void createAverageCountFile(Collection<String> countFiles, String outputFile) {
-		HashMap<String, Integer[]> counts = new HashMap<>();
+		HashMap<String, Double[]> counts = new HashMap<>();
 		LinkedList<String> inputReihenfolge = new LinkedList<>();
 		for (String s : countFiles) {
 			try {
@@ -25,7 +25,7 @@ public class CounterUtils {
 				while ((line = br.readLine()) != null) {
 					split = line.split("\t");
 					inputReihenfolge.add(split[0]);
-					Integer[] countArr = counts.get(split[0]);
+					Double[] countArr = counts.get(split[0]);
 					if (countArr == null) {
 						countArr = parseCounts(split);
 						counts.put(split[0], countArr);
@@ -39,21 +39,22 @@ public class CounterUtils {
 				System.exit(1);
 			}
 		}
-		for (Integer[] countArr : counts.values()) {
+		for (Double[] countArr : counts.values()) {
 			divideArr(countArr, countFiles.size());
 		}
 		writeOutput(outputFile, counts, inputReihenfolge);
 	}
 
-	public static void writeOutput(String outputFile, HashMap<String, Integer[]> counts,
+	public static void writeOutput(String outputFile, HashMap<String, Double[]> counts,
 			LinkedList<String> inputReihenfolge) {
 		try {
 			File f = new File(outputFile);
 			BufferedWriter bw = new BufferedWriter(new FileWriter(f));
 			bw.write("geneId\tnonAmbigousAnyPcr\tnonAmbigousPcr0\tambigousWeightedAnyPcr\tambigousWeightedPcr0\n");
 			for (String s : inputReihenfolge) {
-				Integer[] countArr = counts.get(s);
-				bw.write(s + "\t" + countArr[0] + "\t" + countArr[1] + "\t" + countArr[2] + "\t" + countArr[3] + "\n");
+				Double[] countArr = counts.get(s);
+				bw.write(s + "\t" + countArr[0].intValue() + "\t" + countArr[1].intValue() + "\t" + countArr[2] + "\t"
+						+ countArr[3] + "\n");
 			}
 			bw.close();
 		} catch (Exception e) {
@@ -61,21 +62,21 @@ public class CounterUtils {
 		}
 	}
 
-	public static Integer[] parseCounts(String[] line) {
-		Integer[] ret = new Integer[4];
+	public static Double[] parseCounts(String[] line) {
+		Double[] ret = new Double[4];
 		for (int i = 0; i < 4; i++) {
-			ret[i] = Integer.parseInt(line[i + 1]);
+			ret[i] = (Double.parseDouble(line[i + 1]));
 		}
 		return ret;
 	}
 
-	public static void divideArr(Integer[] arr, int divider) {
+	public static void divideArr(Double[] arr, int divider) {
 		for (int i = 0; i < arr.length; i++) {
 			arr[i] /= divider;
 		}
 	}
 
-	public static void sumIntArray(Integer[] a, Integer[] b) {
+	public static void sumIntArray(Double[] a, Double[] b) {
 		for (int i = 0; i < a.length; i++) {
 			a[i] = a[i] + b[i];
 		}
